refactor(grid_cartopy): move function
Move make_consistent function to grid_helper module.
