+++ goesdr/grid/grid_cartopy.py
@@ -13,10 +13,11 @@ calculate_latlon_grid_cartopy
 """
 
 from netCDF4 import Dataset  # pylint: disable=no-name-in-module
-from numpy import float32, meshgrid, nan, where
+from numpy import float32, meshgrid
 
 from ..projection import GOESABIFixedGridArray, GOESProjection
 from .array import ArrayFloat32, ArrayFloat64
+from .grid_helper import make_consistent
 
 
 def calculate_latlon_grid_cartopy(
@@ -83,19 +84,3 @@ def calculate_latlon_grid_cartopy(
     abi_lon, abi_lat = make_consistent(abi_lon, abi_lat)
 
     return abi_lat.astype(float32), abi_lon.astype(float32)
-
-
-def make_consistent(
-    abi_lon: ArrayFloat64, abi_lat: ArrayFloat64
-) -> tuple[ArrayFloat64, ArrayFloat64]:
-    valid_lon = (abi_lon >= -360.0) & (abi_lon <= 360.0)
-    valid_lat = (abi_lat >= -90.0) & (abi_lat <= 90.0)
-    is_valid = valid_lon & valid_lat
-
-    abi_lon = where(is_valid, abi_lon, nan)
-    abi_lat = where(is_valid, abi_lat, nan)
-
-    abi_lon = where(abi_lon >= 180, abi_lon - 360, abi_lon)
-    abi_lon = where(abi_lon < -180, abi_lon + 360, abi_lon)
-
-    return abi_lon, abi_lat
