+++ goesdr/grid/grid_helper.py
@@ -86,3 +86,38 @@ def compute_latlon_grid(
     abi_lon = where(is_valid, abi_lon, nan)
 
     return abi_lat, abi_lon
+
+
+def make_consistent(
+    abi_lon: ArrayFloat64, abi_lat: ArrayFloat64
+) -> tuple[ArrayFloat64, ArrayFloat64]:
+    """
+    Make the latitude and longitude grids consistent.
+
+    Helper function for functions calculate_latlon_grid_cartopy and
+    calculate_latlon_grid_pyproj.
+
+    Parameters
+    ----------
+    abi_lon : ArrayFloat64
+        The longitude grid data.
+    abi_lat : ArrayFloat64
+        The latitude grid data.
+
+    Returns
+    -------
+    tuple[ArrayFloat64, ArrayFloat64]
+        A tuple containing the consistent latitude and longitude grid
+        data.
+    """
+    valid_lon = (abi_lon >= -360.0) & (abi_lon <= 360.0)
+    valid_lat = (abi_lat >= -90.0) & (abi_lat <= 90.0)
+    is_valid = valid_lon & valid_lat
+
+    abi_lon = where(is_valid, abi_lon, nan)
+    abi_lat = where(is_valid, abi_lat, nan)
+
+    abi_lon = where(abi_lon >= 180, abi_lon - 360, abi_lon)
+    abi_lon = where(abi_lon < -180, abi_lon + 360, abi_lon)
+
+    return abi_lon, abi_lat
